+++ bin/miden-cli/src/commands/tags.rs
@@ -3,7 +3,7 @@ use miden_client::note::{NoteExecutionMode, NoteTag};
 use tracing::info;
 
 use crate::errors::CliError;
-use crate::{Parser, create_dynamic_table, load_config_file};
+use crate::{Parser, create_dynamic_table};
 
 #[derive(Default, Debug, Parser, Clone)]
 #[command(about = "View and manage tags. Defaults to `list` command")]
@@ -41,17 +41,15 @@ impl TagsCmd {
 // HELPERS
 // ================================================================================================
 async fn list_tags<AUTH>(client: Client<AUTH>) -> Result<(), CliError> {
-    let (cli_config, _) = load_config_file()?;
     let mut table = create_dynamic_table(&["Tag", "Source"]);
 
     let tags = client.get_note_tags().await?;
 
     for tag in tags {
         let source = match tag.source {
-            miden_client::sync::NoteTagSource::Account(account_id) => format!(
-                "Account({})",
-                account_id.to_bech32(cli_config.rpc.endpoint.0.to_network_id()?),
-            ),
+            miden_client::sync::NoteTagSource::Account(account_id) => {
+                format!("Account({})", account_id.to_hex())
+            },
             miden_client::sync::NoteTagSource::Note(note_id) => format!("Note({note_id})"),
             miden_client::sync::NoteTagSource::User => "User".to_string(),
         };