+++ crates/rust-client/src/transaction/mod.rs
@@ -620,7 +620,7 @@ where
 
         if ignore_invalid_notes {
             // Remove invalid notes
-            notes = self.get_valid_input_notes(account_id, notes, tx_args.clone()).await?;
+            notes = self.get_valid_input_notes(account, notes, tx_args.clone()).await?;
         }
 
         // Execute the transaction and get the witness
@@ -1054,16 +1054,17 @@ where
 
     async fn get_valid_input_notes(
         &self,
-        account_id: AccountId,
+        account: Account,
         mut input_notes: InputNotes<InputNote>,
         tx_args: TransactionArgs,
     ) -> Result<InputNotes<InputNote>, ClientError> {
         loop {
             let data_store = ClientDataStore::new(self.store.clone());
 
+            data_store.mast_store().load_account_code(account.code());
             let execution = NoteConsumptionChecker::new(&self.build_executor(&data_store)?)
                 .check_notes_consumability(
-                    account_id,
+                    account.id(),
                     self.store.get_sync_height().await?,
                     input_notes.clone(),
                     tx_args.clone(),
@@ -1308,7 +1309,6 @@ mod test {
     use miden_objects::asset::{Asset, FungibleAsset};
     use miden_objects::crypto::dsa::rpo_falcon512::SecretKey;
     use miden_objects::note::NoteType;
-    use miden_objects::testing::account_component::BASIC_WALLET_CODE;
     use miden_objects::testing::account_id::{
         ACCOUNT_ID_PRIVATE_FUNGIBLE_FAUCET,
         ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET,
@@ -1337,7 +1337,10 @@ mod test {
         keystore.add_key(&AuthSecretKey::RpoFalcon512(secret_key)).unwrap();
 
         let wallet_component = AccountComponent::compile(
-            BASIC_WALLET_CODE,
+            "
+                export.::miden::contracts::wallets::basic::receive_asset
+                export.::miden::contracts::wallets::basic::move_asset_to_note
+            ",
             TransactionKernel::assembler(),
             vec![StorageSlot::Value(Word::default()), StorageSlot::Map(StorageMap::default())],
         )